seems great

+++ include/lem_in.h
@@ -6,7 +6,7 @@
 /*   By: conoel <conoel@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/16 18:53:12 by conoel            #+#    #+#             */
-/*   Updated: 2019/04/01 17:03:04 by conoel           ###   ########.fr       */
+/*   Updated: 2019/04/02 15:52:48 by conoel           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,8 +38,8 @@ typedef struct		s_path
 /*
 **	PARSING FUNCTIONS
 */
-t_node		*load_map(int argc, char ** argv, int *ant_nb);
-t_node		*load_structure(char *data, int *ant_nb);
+t_node		*load_map(int argc, char ** argv, long *ant_nb);
+t_node		*load_structure(char *data, long *ant_nb);
 t_node		*add_node(t_node *head, char *line, int pos);
 void		free_nodes(t_node *head);
 t_node		*get_node(char *name, size_t len, t_node *head);
